restructure bundler package
